requirements.txt
asgiref==3.8.1
diff-match-patch==20241021
Django==3.1.12
django-import-export==4.3.7
djongo==1.3.7
dnspython==2.7.0
et_xmlfile==2.0.0
gunicorn==23.0.0
openpyxl==3.1.5
packaging==24.2
pymongo==3.11.4
python-dotenv==1.1.0
pytz==2025.2
sqlparse==0.2.4
tablib==3.8.0
tzdata==2025.2
